Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 16ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240127/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.cpp
@@ -4,11 +4,11 @@
 using namespace std;
 
 long long m;
-long long BinarySearch(int target) {
+long long BinarySearch(long long target) {
 	long long left = 0;
 	long long right = m-1;
 	long long mid = 0;
-	int cnt = 0;
+	long long cnt = 0;
     target--;
 
 	while (left <= right) {
@@ -33,7 +33,7 @@ int main() {
 	cin >> m >> a >> b;
 	long long small = LLONG_MAX;
 	long long big = 0;
-	for (int i = a; i <= b; i++) {
+	for (long long i = a; i <= b; i++) {
 		long long num = BinarySearch(i);
 		small = min(small, num);
 		big = max(big, num);
